[CHORE] PortfolioService 추상화
